+++ CHANGELOG.md
@@ -33,7 +33,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
-- Fix regression where `--playground-tab-bar-background` and
+- Playground can now be imported from Skypack.
+
+- Fix bugs where `--playground-tab-bar-background` and
   `--playground-tab-bar-foreground-color` did not apply correctly.
 
 ## [0.5.0] - 2021-01-30